Adapt Test suites to use registerBinding
Previously, registerBinding did not produce decorated elements in
a polymorphic container. This made it impossible to use this to
register types that had managed properties.

To make use of registerBinding, the JvmTestSuite and JvmTestSuiteTarget
constructors must only take a name and registered services.

This change only relies on available services and moves other configuration
into the plugin